Update addr@ code style, move get method to deployer
