+++ src/toncli/modules/abstract/deployer.py
@@ -1,23 +1,43 @@
 import os
 import sys
-from itertools import cycle
-from time import sleep
-from typing import Tuple, Optional, List
+import tempfile
+from argparse import Namespace
+from typing import List
 
 from colorama import Fore, Style
+from itertools import cycle
+from time import sleep
+from typing import Tuple
 
-from toncli.modules.utils.system.conf import project_root
 from toncli.modules.utils.fift.commands import contract_manipulation, test_fift
 from toncli.modules.utils.func.commands import build as fift_build
 from toncli.modules.utils.lite_client.commands import get_account_status, send_boc
 from toncli.modules.utils.system.project import check_for_needed_files_to_deploy
-from toncli.modules.utils.lite_client.lite_client import LiteClient
+from toncli.modules.utils.system.project_conf import TonProjectConfig
 from toncli.modules.utils.system.log import logger
-from toncli.modules.utils.system.project_conf import ProjectConf, TonProjectConfig
+from toncli.modules.utils.fift.fift import Fift
+from toncli.modules.utils.system.project_conf import ProjectConf
+from toncli.modules.utils.ton.cell import deserialize
+from toncli.modules.utils.lite_client.parser import split_get_output
+from jinja2 import FileSystemLoader, Environment, select_autoescape
+
+from toncli.modules.utils.lite_client.lite_client import LiteClient
+from toncli.modules.utils.system.conf import project_root
 
 bl = Fore.CYAN
 gr = Fore.GREEN
 rs = Style.RESET_ALL
+
+
+class Cell:
+    def __init__(self, data: str, refs: List['Cell']):
+        self.data = data
+        self.refs = refs
+
+    def serialize(self):
+        refs = [cell.serialize() for cell in self.refs]
+        refs = [f"{ref} ref," for ref in refs]
+        return f"<b {self.data} s, {' '.join(refs)} b>"
 
 
 class AbstractDeployer:
@@ -182,3 +202,140 @@ class AbstractDeployer:
             seqnos.append(output)
 
         return seqnos
+
+    def get(self, args: List[str], kwargs: Namespace):
+        """Run get methods on contracts"""
+        # TODO: it wasn't good idea to parse lite client output
+        # We can run get methods locally by runvm (savedata / saveaccount) and then get output / run fift
+        # It's needed to be done this way
+
+        contracts = kwargs.contracts.split() if kwargs.contracts else None
+        if contracts is not None and len(contracts) > 0:
+            real_contracts = []
+
+            for item in contracts:
+                for config in self.project_config.contracts:
+                    if config.name == item:
+                        real_contracts.append(config)
+        else:
+            real_contracts = self.project_config.contracts
+
+        # Get contracts addresses
+        self.addresses = self.get_address(real_contracts)
+
+        for address, contract in zip(self.addresses, real_contracts):
+            logger.info(f"👯 [{bl}{contract.name}{rs}] [{gr}{address[1]}{rs}] runmethod {args}")
+            lite_client = LiteClient('runmethod', args=[address[1], *args], kwargs={'lite_client_args': '-v 0',
+                                                                                    'net': self.network,
+                                                                                    'update': self.update_config},
+                                     get_output=True)
+            output = lite_client.run_safe()
+            output = output.split('\n')[-3]
+            output = output[11:-2]
+            logger.info(f"🧐 Output: [ {output} ]")
+            if kwargs.fift and len(kwargs.fift) > 0:
+                output = split_get_output(output)
+                to_fift = []
+
+                # TODO: use fift runvm
+                # This is not right, but we have not time
+                # This code is load C{...} from lite client to fift code
+                for line in output:
+                    # If cell hash present
+                    if line[:2] == 'C{':
+                        _hash = line[2:-1]
+
+                        # we need to parse cell
+                        lite_client = LiteClient('runmethod', args=[address[1], *args],
+                                                 kwargs={'lite_client_args': '-v 0',
+                                                         'net': self.network,
+                                                         'update': self.update_config,
+                                                         'lite_client_post_args': ["-c", f"dumpcell {_hash}"]},
+                                                 get_output=True)
+                        output_cells = lite_client.run_safe()
+
+                        cells = []
+
+                        append_other = False
+                        # cell can contains references
+                        for line in output_cells.split('\n'):
+                            if append_other and len(line):
+                                cells.append(line[2:])
+
+                            if _hash.upper() in line and '} =' in line:
+                                append_other = True
+
+                        main_cell = None
+                        for cell in cells:
+                            level = len(cell) - len(cell.strip())
+
+                            if level == 0:
+                                main_cell = Cell(data=cell, refs=[])
+                            else:
+                                needed_cell: Cell = None
+
+                                for current_level in range(level):
+                                    if not needed_cell:
+                                        needed_cell = main_cell
+                                    else:
+                                        needed_cell = needed_cell.refs[-1]
+
+                                needed_cell.refs.append(Cell(data=cell.strip(), refs=[]))
+                        to_fift.append(main_cell.serialize())
+                    elif line[:3] == 'CS{':
+                        # Дай бог здоровья больше не писать такой код
+                        # И использовать fift runvm
+                        # Но к несчастью я не догадался так сделать сначала
+                        # А сейчас хочется сделать продакшн реди как можно быстрее
+                        # Если кто-то это переделает - буду очень благодарен
+                        # disintar.io
+
+                        cell = line.split()[0].replace('CS{Cell{', '')[:-1]
+
+                        cut = list(map(int, line.split()[2][:-1].split('..')))
+
+                        bits = deserialize(cell, *cut)
+
+                        bits = ["<b b{%s} s, b> <s" % bits[i:i + 128] for i in range(0, len(bits), 128)]
+                        if len(bits) > 1:
+                            bits = f'{bits[0]} {" |+ ".join(bits[1:])} |+'
+                        else:
+                            bits = bits[0]
+
+                        to_fift.append(bits)
+                    else:
+                        to_fift.append(line)
+
+                render_kwargs = {
+                    'code': kwargs.fift,
+                    'output': ' '.join(to_fift)
+                }
+
+                # Load template of transaction_debug
+                loader = FileSystemLoader(f"{project_root}/modules/fift")
+
+                env = Environment(
+                    loader=loader,
+                    autoescape=select_autoescape()
+                )
+
+                template = env.get_template("get_run.fif.template")
+
+                rendered = template.render(**render_kwargs)
+
+                temp_location: str = tempfile.mkstemp(suffix='.fif')[1]
+
+                with open(temp_location, 'w') as f:
+                    f.write(rendered)
+
+                fift = Fift('run', args=[temp_location])
+                fift.run()
+            else:
+                try:
+                    output = int(output)
+                    hex = "{0:x}".format(output)
+
+                    string_output = bytearray.fromhex(hex).decode()
+                    logger.info(f"🧐 Auto parse string: [ {string_output} ]")
+                except Exception as e:
+                    logger.error(f"🧐 Can't auto parse string")
